refactor: move getFormattedTime to time.ts will enable better reuse of this fxn
