+++ packages/shared/utils/time.ts
@@ -80,3 +80,29 @@ export function toPointRange(range: TimeStampedPointRange | PointRange): PointRa
     };
   }
 }
+
+// Format a time value to mm:ss
+export function getFormattedTime(time: number, showMilliseconds: boolean = false) {
+  const date = new Date(time);
+  let minutes = date.getUTCMinutes();
+  let seconds = date.getUTCSeconds();
+  const milliseconds = date.getUTCMilliseconds();
+
+  if (!showMilliseconds) {
+    if (milliseconds >= 500) {
+      seconds++;
+    }
+    if (seconds >= 60) {
+      seconds = 0;
+      minutes++;
+    }
+  }
+
+  if (showMilliseconds) {
+    return `${minutes}:${seconds.toString().padStart(2, "0")}.${milliseconds
+      .toString()
+      .padStart(3, "0")}`;
+  } else {
+    return `${minutes}:${seconds.toString().padStart(2, "0")}`;
+  }
+}
