+++ src/ui/utils/timeline.ts
@@ -120,32 +120,6 @@ export function getNewZoomRegion({
   return { begin: newBegin, end: newEnd };
 }
 
-// Format a time value to mm:ss
-export function getFormattedTime(time: number, showMilliseconds: boolean = false) {
-  const date = new Date(time);
-  let minutes = date.getUTCMinutes();
-  let seconds = date.getUTCSeconds();
-  const milliseconds = date.getUTCMilliseconds();
-
-  if (!showMilliseconds) {
-    if (milliseconds >= 500) {
-      seconds++;
-    }
-    if (seconds >= 60) {
-      seconds = 0;
-      minutes++;
-    }
-  }
-
-  if (showMilliseconds) {
-    return `${minutes}:${seconds.toString().padStart(2, "0")}.${milliseconds
-      .toString()
-      .padStart(3, "0")}`;
-  } else {
-    return `${minutes}:${seconds.toString().padStart(2, "0")}`;
-  }
-}
-
 const TIME_REGEX = /^([0-9]+)(:[0-9]{0,2}){0,1}(\.[0-9]{0,3}){0,1}$/;
 
 export function isValidTimeString(formatted: string): boolean {
